Use `off_t` instead of `off64_t` for musl compatibility.

+++ src/test/proc_mem.c
@@ -131,15 +131,15 @@ static void do_test(int (*opener)(int)) {
   fd = opener(child);
   test_assert(fd >= 0);
   test_assert(sizeof(COOKIE) ==
-              pwrite(fd, &COOKIE, sizeof(COOKIE), (off_t)&cookie1));
+              pwrite(fd, &COOKIE, sizeof(COOKIE), (uintptr_t)&cookie1));
 
   iov[0].iov_base = (char*)&COOKIE;
   iov[0].iov_len = 2;
   iov[1].iov_base = (char*)&COOKIE + 2;
   iov[1].iov_len = 2;
-  test_assert(sizeof(COOKIE) == pwritev(fd, iov, 2, (off_t)&cookie2));
+  test_assert(sizeof(COOKIE) == pwritev(fd, iov, 2, (uintptr_t)&cookie2));
 
-  lseek(fd, (off_t)&cookie3, SEEK_SET);
+  lseek(fd, (uintptr_t)&cookie3, SEEK_SET);
   test_assert(sizeof(COOKIE) == write(fd, &COOKIE, sizeof(COOKIE)));
 
   test_assert(1 == write(pipe_fds[1], "x", 1));
